Move is_not into specialize, some experiments with the macros.
