+++ src/dev_utilities/mod.rs
@@ -1,6 +1,6 @@
 //! Utilities to be used when developing backends and middleware.
 pub mod vecstorage;
-pub mod is_not;
+#[macro_use]
 pub mod specialize;
 #[macro_use]
 pub mod compatibility;
